raftstore/store: move check_epoch from peer to util
